Create player_info.py
Takes '?username=WHATEVER' as input.  The UserName field of the database shold always be unique, if not, or if there are no entries for the given username, this page will display an error with how many results were returned.  If only one result was returned the page will be structured:

Name: First Last (a.k.a. UserName)
Role: role
Email: email
profile description

as an html table with colspan = 2 for profile description.